+++ docs/ai/stories-guide.md
@@ -88,25 +88,20 @@ stories/
 
 Если файл сторей компонента находится в папке `stories`, необходимо создать одноименную папку и перенести файл сторей туда.
 
-Все стори, кроме **Visual tests**, выносятся в подпапку `examples/` и импортируются в основной файл. Если в основной папке находятся несколько файлов с расширением `.stories.tsx`, внутри подпапки `examples` необходимо создать подпапки по названию компонентов из файлов с расширением `.stories.tsx`. Также необходимо в каждой такой подпапке создать отдельный файл `index.ts`.
-
-Playground и Visual tests остаются inline в файле stories, потому что:
-
-- **Playground** — интерактивная песочница, не предназначена для копирования.
-- **Visual tests** — тестовый код (захардкоженные данные, `isOpen`, `play`-функции), не предназначен для копирования. Не показывает исходный код в docs, поэтому `?raw`-экспорт не нужен.
+Все стори выносятся в подпапку `examples/` и импортируются в основной файл. Если в основной папке находятся несколько файлов с расширением `.stories.tsx`, внутри подпапки `examples` необходимо создать подпапки по названию компонентов из файлов с расширением `.stories.tsx`. Также необходимо в каждой такой подпапке создать отдельный файл `index.ts`.
 
 ---
 
 ## Обязательные стори
 
-| Стори                         | Описание                                                                                                                              | Controls | Код |
-| ----------------------------- | ------------------------------------------------------------------------------------------------------------------------------------- | -------- | --- |
-| **Playground**                | Интерактивная песочница с Controls                                                                                                    | Да       | Нет |
-| **Default**                   | Минимальное состояние компонента с параметрами по умолчанию                                                                           | Нет      | Да  |
-| **Sizes / Themes / Statuses** | Стори на ключевые props — отрендеренные варианты с подписями                                                                          | Нет      | Да  |
-| **Edge cases**                | Опционально, на усмотрение разработчика. Каждый кейс — отдельная стори                                                                | Нет      | Да  |
-| **Examples**                  | Композиции компонентов, production-like примеры                                                                                       | Нет      | Да  |
-| **Visual tests**              | Дополнительные примеры для скриншот-тестов (если недостаточно основных); исходный код в docs не показывается, `?raw`-экспорт не нужен | Нет      | Нет |
+| Стори                         | Описание                                                                                                     | Controls | Код |
+| ----------------------------- | ------------------------------------------------------------------------------------------------------------ | -------- | --- |
+| **Playground**                | Интерактивная песочница с Controls                                                                           | Да       | Нет |
+| **Default**                   | Минимальное состояние компонента с параметрами по умолчанию                                                  | Нет      | Да  |
+| **Sizes / Themes / Statuses** | Стори на ключевые props — отрендеренные варианты с подписями                                                 | Нет      | Да  |
+| **Edge cases**                | Опционально, на усмотрение разработчика. Каждый кейс — отдельная стори                                       | Нет      | Да  |
+| **Examples**                  | Композиции компонентов, production-like примеры                                                              | Нет      | Да  |
+| **Visual tests**              | Дополнительные примеры для скриншот-тестов (если недостаточно основных); исходный код в docs не показывается | Нет      | Нет |
 
 > **Правило:** только Playground имеет Controls. Документационные стори (`Default`, `Sizes / Themes / Statuses`, `Edge cases`, `Examples`) не имеют Controls и показывают пример кода. `Visual tests` не имеют Controls и не показывают код.
 
@@ -399,7 +394,7 @@ export { default as SizesExampleSource } from "./SizesExample?raw";
 
 ## Подключение примеров к стори
 
-В modern pattern каждая документационная стори (кроме Visual tests) подключает пример и его исходный код через `?raw`:
+В modern pattern каждая документационная стори подключает пример и его исходный код через `?raw`:
 
 ```tsx
 export const Default: StoryObj<typeof Component> = {
