+++ stories/Amount/examples/index.ts
@@ -2,6 +2,5 @@ export * from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export * from "./ShowcaseExample";
 export { default as ShowcaseExampleSource } from "./ShowcaseExample?raw";
-
 export * from "./PlaygroundExample";
 export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
