+++ stories/Buttons/ButtonDropdown.stories.tsx
@@ -1,14 +1,6 @@
-import React, { useMemo, useState } from "react";
+import React from "react";
 import { Meta, StoryObj } from "@storybook/react";
-import { action } from "storybook/actions";
-import {
-    ButtonDropdown,
-    EButtonDotsTheme,
-    EButtonTheme,
-    EComponentSize,
-    EScreenWidth,
-    IButtonDropdownOption,
-} from "@sberbusiness/triplex-next";
+import { ButtonDropdown, EButtonDotsTheme, EButtonTheme, EComponentSize } from "@sberbusiness/triplex-next";
 import { Title, Description, Primary, Controls, Stories, ArgTypes, Heading } from "@storybook/addon-docs/blocks";
 import {
     BlockModeExample,
@@ -24,35 +16,8 @@ import {
     ThemesExampleSource,
     WithSelectedOptionExample,
     WithSelectedOptionExampleSource,
+    VisualTestsExample,
 } from "./examples/ButtonDropdown/index";
-import { AdaptiveUtils } from "../utils/adaptiveUtils";
-
-const createOptions = (onItemSelect?: (id: string) => void): IButtonDropdownOption[] => [
-    {
-        id: "opt-1",
-        label: "Действие 1",
-        onSelect: () => {
-            action("select")("opt-1");
-            onItemSelect?.("opt-1");
-        },
-    },
-    {
-        id: "opt-2",
-        label: "Действие 2",
-        onSelect: () => {
-            action("select")("opt-2");
-            onItemSelect?.("opt-2");
-        },
-    },
-    {
-        id: "opt-3",
-        label: "Действие 3",
-        onSelect: () => {
-            action("select")("opt-3");
-            onItemSelect?.("opt-3");
-        },
-    },
-];
 
 const meta = {
     title: "Components/Buttons/ButtonDropdown",
@@ -258,45 +223,7 @@ export const WithSelectedOption: Story = {
 
 export const VisualTests: Story = {
     tags: ["!autodocs"],
-    render: () => {
-        const [selectedId, setSelectedId] = useState<string | undefined>("opt-2");
-        const options = useMemo(() => createOptions(setSelectedId), []);
-        const selectedSM = options.find((o) => o.id === selectedId);
-        const selectedMD = options.find((o) => o.id === selectedId);
-        const selectedLG = options.find((o) => o.id === selectedId);
-
-        return (
-            <div style={{ display: "flex", gap: 16 }}>
-                <ButtonDropdown
-                    opened={true}
-                    theme={EButtonTheme.GENERAL}
-                    size={EComponentSize.SM}
-                    options={options}
-                    selected={selectedSM}
-                >
-                    Button text
-                </ButtonDropdown>
-                <ButtonDropdown
-                    opened={!AdaptiveUtils.isAdaptive(EScreenWidth.SM_MAX)}
-                    theme={EButtonTheme.GENERAL}
-                    size={EComponentSize.MD}
-                    options={options}
-                    selected={selectedMD}
-                >
-                    Button text
-                </ButtonDropdown>
-                <ButtonDropdown
-                    opened={!AdaptiveUtils.isAdaptive(EScreenWidth.SM_MAX)}
-                    theme={EButtonTheme.GENERAL}
-                    size={EComponentSize.LG}
-                    options={options}
-                    selected={selectedLG}
-                >
-                    Button text
-                </ButtonDropdown>
-            </div>
-        );
-    },
+    render: VisualTestsExample,
     parameters: {
         controls: { disable: true },
         docs: {
