+++ stories/Chips/examples/ChipSort/index.ts
@@ -4,3 +4,4 @@ export { SizesExample } from "./SizesExample";
 export { default as SizesExampleSource } from "./SizesExample?raw";
 export { PlaygroundExample, type ChipSortPlaygroundArgs } from "./PlaygroundExample";
 export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
+export { VisualTestsExample } from "./VisualTestsExample";
