+++ stories/Chips/examples/ChipSuggest/index.ts
@@ -8,3 +8,4 @@ export { WithNotificationIconExample } from "./WithNotificationIconExample";
 export { default as WithNotificationIconExampleSource } from "./WithNotificationIconExample?raw";
 export { PlaygroundExample } from "./PlaygroundExample";
 export { default as PlaygroundExampleSource } from "./PlaygroundExample?raw";
+export { VisualTestsExample } from "./VisualTestsExample";
