+++ stories/Col/examples/index.ts
@@ -1,3 +1,4 @@
+export { PlaygroundExample } from "./PlaygroundExample";
 export { DefaultExample } from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export { DifferentSizesExample } from "./DifferentSizesExample";
