+++ stories/DateField/examples/index.ts
@@ -1,6 +1,8 @@
+export { PlaygroundExample } from "./PlaygroundExample";
 export { DefaultExample } from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export { WithPostfixExample } from "./WithPostfixExample";
 export { default as WithPostfixExampleSource } from "./WithPostfixExample?raw";
 export { StatesExample } from "./StatesExample";
 export { default as StatesExampleSource } from "./StatesExample?raw";
+export { VisualTestsExample } from "./VisualTestsExample";
