+++ stories/DateRange/examples/index.ts
@@ -1,4 +1,6 @@
+export { PlaygroundExample } from "./PlaygroundExample";
 export { DefaultExample } from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export { WithoutNavigationExample } from "./WithoutNavigationExample";
 export { default as WithoutNavigationExampleSource } from "./WithoutNavigationExample?raw";
+export { VisualTestsExample } from "./VisualTestsExample";
