+++ stories/DocumentNumberEdit/examples/index.ts
@@ -1,2 +1,4 @@
+export { PlaygroundExample } from "./PlaygroundExample";
 export { DefaultExample } from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
+export { VisualTestsExample } from "./VisualTestsExample";
