+++ stories/Ellipsis/examples/index.ts
@@ -1,2 +1,3 @@
+export { PlaygroundExample } from "./PlaygroundExample";
 export { DefaultExample } from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
