+++ stories/FormField/FormField.stories.tsx
@@ -1,18 +1,6 @@
-import React, { useState } from "react";
+import React from "react";
 import { Meta, StoryObj } from "@storybook/react";
-import {
-    EComponentSize,
-    EFontType,
-    EFormFieldStatus,
-    ETextSize,
-    FormField,
-    FormFieldDescription,
-    FormFieldInput,
-    FormFieldLabel,
-    FormFieldMaskedInput,
-    FormGroup,
-    Text,
-} from "@sberbusiness/triplex-next";
+import { EComponentSize, EFormFieldStatus, FormField } from "@sberbusiness/triplex-next";
 import {
     Title as DocsTitle,
     Description,
@@ -23,6 +11,8 @@ import {
     ArgTypes,
 } from "@storybook/addon-docs/blocks";
 import {
+    PlaygroundExample,
+    type IFormFieldPlaygroundProps,
     DefaultExample,
     DefaultExampleSource,
     MaskedInputExample,
@@ -39,6 +29,7 @@ import {
     WithCounterExampleSource,
     WithPrefixAndPostfixExample,
     WithPrefixAndPostfixExampleSource,
+    VisualTestsExample,
 } from "./examples";
 
 const meta = {
@@ -68,44 +59,12 @@ const meta = {
 
 export default meta;
 
-interface IFormFieldPlaygroundProps extends React.ComponentProps<typeof FormField> {
-    labelText?: string;
-    placeholder?: string;
-    showClear?: boolean;
-    descriptionText?: string;
-    counter?: string;
-}
-
 type Story = StoryObj<typeof FormField>;
 type PlaygroundStory = StoryObj<IFormFieldPlaygroundProps>;
 
 export const Playground: PlaygroundStory = {
     tags: ["!autodocs"],
-    render: function Render(args) {
-        const [value, setValue] = useState("");
-        const { labelText, placeholder, descriptionText, counter, ...formFieldProps } = args;
-        return (
-            <div style={{ maxWidth: "300px" }}>
-                <FormGroup>
-                    <FormField {...formFieldProps}>
-                        <FormFieldLabel>{labelText || "Название поля"}</FormFieldLabel>
-                        <FormFieldInput
-                            value={value}
-                            onChange={(e) => setValue(e.target.value)}
-                            placeholder={placeholder || "Введите текст..."}
-                        />
-                    </FormField>
-                    {(descriptionText || counter) && (
-                        <FormFieldDescription>
-                            <Text tag="div" size={ETextSize.B4} type={EFontType.SECONDARY}>
-                                {descriptionText || "Описание поля"}
-                            </Text>
-                        </FormFieldDescription>
-                    )}
-                </FormGroup>
-            </div>
-        );
-    },
+    render: PlaygroundExample,
     argTypes: {
         status: {
             control: { type: "select" },
@@ -171,68 +130,8 @@ export const MaskedInput: Story = {
     parameters: { controls: { disable: true }, docs: { source: { code: MaskedInputExampleSource, language: "tsx" } } },
 };
 
-export const VisualTests: StoryObj<typeof FormFieldMaskedInput> = {
+export const VisualTests: StoryObj<typeof FormField> = {
     tags: ["!autodocs"],
-    render: function Render() {
-        return (
-            <div style={{ display: "flex", flexDirection: "column", gap: 24 }}>
-                <div style={{ display: "flex", gap: 24 }}>
-                    <FormGroup>
-                        <FormField size={EComponentSize.SM}>
-                            <FormFieldLabel>Label</FormFieldLabel>
-                            <FormFieldMaskedInput
-                                value="1234 5678 9012 3456"
-                                onChange={() => {}}
-                                mask={FormFieldMaskedInput.presets.masks.cardNumber}
-                                placeholderMask={FormFieldMaskedInput.presets.placeholderMasks.cardNumber}
-                            />
-                        </FormField>
-                    </FormGroup>
-                    <FormGroup>
-                        <FormField size={EComponentSize.MD}>
-                            <FormFieldLabel>Label</FormFieldLabel>
-                            <FormFieldMaskedInput
-                                value="1234 5678 9012 3456"
-                                onChange={() => {}}
-                                mask={FormFieldMaskedInput.presets.masks.cardNumber}
-                                placeholderMask={FormFieldMaskedInput.presets.placeholderMasks.cardNumber}
-                            />
-                        </FormField>
-                    </FormGroup>
-                    <FormGroup>
-                        <FormField size={EComponentSize.LG}>
-                            <FormFieldLabel>Label</FormFieldLabel>
-                            <FormFieldMaskedInput
-                                value="1234 5678 9012 3456"
-                                onChange={() => {}}
-                                mask={FormFieldMaskedInput.presets.masks.cardNumber}
-                                placeholderMask={FormFieldMaskedInput.presets.placeholderMasks.cardNumber}
-                            />
-                        </FormField>
-                    </FormGroup>
-                </div>
-                <div style={{ display: "flex", gap: 24 }}>
-                    <FormGroup>
-                        <FormField size={EComponentSize.SM} active>
-                            <FormFieldLabel>Label</FormFieldLabel>
-                            <FormFieldInput value="" onChange={() => {}} placeholder="Введите текст..." />
-                        </FormField>
-                    </FormGroup>
-                    <FormGroup>
-                        <FormField size={EComponentSize.MD} active>
-                            <FormFieldLabel>Label</FormFieldLabel>
-                            <FormFieldInput value="" onChange={() => {}} placeholder="Введите текст..." />
-                        </FormField>
-                    </FormGroup>
-                    <FormGroup>
-                        <FormField size={EComponentSize.LG} active>
-                            <FormFieldLabel>Label</FormFieldLabel>
-                            <FormFieldInput value="" onChange={() => {}} placeholder="Введите текст..." />
-                        </FormField>
-                    </FormGroup>
-                </div>
-            </div>
-        );
-    },
+    render: VisualTestsExample,
     parameters: { controls: { disable: true }, docs: { canvas: { sourceState: "none" }, codePanel: false } },
 };
