+++ stories/FormField/examples/index.ts
@@ -1,3 +1,4 @@
+export { PlaygroundExample, type IFormFieldPlaygroundProps } from "./PlaygroundExample";
 export { DefaultExample } from "./DefaultExample";
 export { default as DefaultExampleSource } from "./DefaultExample?raw";
 export { WithPrefixAndPostfixExample } from "./WithPrefixAndPostfixExample";
@@ -14,3 +15,4 @@ export { SizesExample } from "./SizesExample";
 export { default as SizesExampleSource } from "./SizesExample?raw";
 export { MaskedInputExample } from "./MaskedInputExample";
 export { default as MaskedInputExampleSource } from "./MaskedInputExample?raw";
+export { VisualTestsExample } from "./VisualTestsExample";
